runner refactor phase 1
